requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Cython==0.29.1
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10
joblib==0.13.0
lightfm==1.15
MarkupSafe==1.1.0
numpy==1.15.4
pandas==0.23.4
pyhanlp==0.1.44
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
scikit-surprise==1.0.6
scipy==1.2.0rc2
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
